+++ src/player/movement.rs
@@ -1,300 +1,5 @@
 //! # Player's Movement controlling systems
 
-use crate::player::Player;
-use avian3d::prelude::*;
-use bevy::{input::mouse::AccumulatedMouseMotion, prelude::*};
-
-pub fn keyboard_mouse_system(
-    mut commands: Commands,
-    mut query: Query<
-        (
-            &mut Transform,
-            &mut AngularVelocity,
-            &mut LinearVelocity,
-            &Player,
-        ),
-        With<Player>,
-    >,
-    accumulated_mouse_motion: Res<AccumulatedMouseMotion>,
-    keyboard: Res<ButtonInput<KeyCode>>,
-    asset_server: Res<AssetServer>,
-) {
-    let (transform, mut angular, mut linear, player) = query.single_mut().unwrap();
-
-    // Mouse control
-    if accumulated_mouse_motion.delta != Vec2::ZERO {
-        let delta = accumulated_mouse_motion.delta;
-
-        let x = delta.x * -1. / 100.;
-        let y = delta.y * -1. / 100.;
-        let mouse: Vec2 = Vec2::new(x, y);
-
-        let rotation: Quat = transform.rotation;
-
-        let mut velocity: Vec3 = Vec3::ZERO;
-
-        // Mouse X
-        {
-            let direction: Vec3 = if player.config.mouse.x_reverse {
-                rotation * Vec3::NEG_Y
-            } else {
-                rotation * Vec3::Y
-            };
-            // dbg!(direction);
-
-            let x_result: f32 = mouse.x * direction.x;
-            let y_result: f32 = mouse.x * direction.y;
-            let z_result: f32 = mouse.x * direction.z;
-
-            velocity.x += x_result;
-            velocity.y += y_result;
-            velocity.z += z_result;
-        }
-
-        // Mouse Y
-        {
-            let direction: Vec3 = if player.config.mouse.y_reverse {
-                rotation * Vec3::NEG_X
-            } else {
-                rotation * Vec3::X
-            };
-            // dbg!(direction);
-
-            let x_result: f32 = mouse.y * direction.x;
-            let y_result: f32 = mouse.y * direction.y;
-            let z_result: f32 = mouse.y * direction.z;
-
-            velocity.x += x_result;
-            velocity.y += y_result;
-            velocity.z += z_result;
-        }
-
-        angular.x += velocity.x;
-        angular.y += velocity.y;
-        angular.z += velocity.z;
-    }
-
-    // Hovering
-    if keyboard.just_pressed(player.config.keyboard.hover) {
-        angular.0 = Vec3::ZERO;
-        linear.0 = Vec3::ZERO;
-    }
-
-    // Moving
-    if keyboard.just_pressed(player.config.keyboard.forward) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        let mut velocity: Vec3 = Vec3::ZERO;
-        const FORCE: f32 = 10.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::NEG_Z;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        velocity += result;
-
-        linear.0 += velocity;
-    }
-
-    if keyboard.just_pressed(player.config.keyboard.left) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        let mut velocity: Vec3 = Vec3::ZERO;
-        const FORCE: f32 = 10.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::NEG_X;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        velocity += result;
-
-        linear.0 += velocity;
-    }
-
-    if keyboard.just_pressed(player.config.keyboard.back) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        let mut velocity: Vec3 = Vec3::ZERO;
-        const FORCE: f32 = 10.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::Z;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        velocity += result;
-
-        linear.0 += velocity;
-    }
-
-    if keyboard.just_pressed(player.config.keyboard.right) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        let mut velocity: Vec3 = Vec3::ZERO;
-        const FORCE: f32 = 10.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::X;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        velocity += result;
-
-        linear.0 += velocity;
-    }
-
-    if keyboard.just_pressed(player.config.keyboard.roll_left) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        const FORCE: f32 = 3.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::NEG_Z;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        angular.0 += result;
-    }
-
-    if keyboard.just_pressed(player.config.keyboard.roll_right) {
-        commands.spawn((
-            AudioPlayer::new(asset_server.load("SE/engine_dash.ogg")),
-            PlaybackSettings::ONCE.with_spatial(false),
-        ));
-
-        const FORCE: f32 = 3.0;
-
-        let rotation: Quat = transform.rotation;
-        let direction: Vec3 = rotation * Vec3::Z;
-
-        let x: f32 = FORCE * direction.x;
-        let y: f32 = FORCE * direction.y;
-        let z: f32 = FORCE * direction.z;
-        let result: Vec3 = Vec3::new(x, y, z);
-
-        angular.0 += result;
-    }
-}
-
-pub fn controller_system(
-    mut query: Query<(&mut Transform, &mut AngularVelocity, &mut LinearVelocity), With<Player>>,
-    gamepads: Query<(Entity, &Gamepad)>,
-) {
-    for (entity, gamepad) in &gamepads {
-        let left_stick = gamepad.left_stick();
-        let right_stick = gamepad.right_stick();
-
-        debug!("{} right_stick: {}", entity, right_stick);
-        debug!("{} left_stick: {}", entity, left_stick);
-
-        // Hovering
-        if gamepad.just_pressed(GamepadButton::DPadDown) {
-            for (_transform, mut angular, mut linear) in &mut query {
-                angular.0 = Vec3::ZERO;
-                linear.0 = Vec3::ZERO;
-            }
-        }
-
-        // Moving
-        for (transform, _angular, mut linear) in &mut query {
-            let rotation: Quat = transform.rotation;
-            let mut velocity: Vec3 = Vec3::ZERO;
-
-            // LeftStick X
-            if left_stick.x.abs() > 0.1 {
-                let direction: Vec3 = rotation * Vec3::X;
-
-                let x_result: f32 = left_stick.x * direction.x;
-                let y_result: f32 = left_stick.x * direction.y;
-                let z_result: f32 = left_stick.x * direction.z;
-
-                velocity.x += x_result;
-                velocity.y += y_result;
-                velocity.z += z_result;
-            }
-
-            // LeftStick Y
-            if left_stick.y.abs() > 0.1 {
-                let direction: Vec3 = rotation * Vec3::NEG_Z;
-
-                let x_result: f32 = left_stick.y * direction.x;
-                let y_result: f32 = left_stick.y * direction.y;
-                let z_result: f32 = left_stick.y * direction.z;
-
-                velocity.x += x_result;
-                velocity.y += y_result;
-                velocity.z += z_result;
-            }
-
-            linear.0 += velocity;
-        }
-
-        for (transform, mut angular, mut _linear) in &mut query {
-            let rotation: Quat = transform.rotation;
-
-            let mut velocity: Vec3 = Vec3::ZERO;
-
-            // RightStick X
-            if right_stick.x.abs() > 0.1 {
-                let direction: Vec3 = rotation * Vec3::Y;
-
-                let x_result: f32 = -right_stick.x * direction.x;
-                let y_result: f32 = -right_stick.x * direction.y;
-                let z_result: f32 = -right_stick.x * direction.z;
-
-                velocity.x += x_result;
-                velocity.y += y_result;
-                velocity.z += z_result;
-            }
-
-            // RightStick Y
-            if right_stick.y.abs() > 0.1 {
-                let direction: Vec3 = rotation * Vec3::X;
-
-                let x_result: f32 = right_stick.y * direction.x;
-                let y_result: f32 = right_stick.y * direction.y;
-                let z_result: f32 = right_stick.y * direction.z;
-
-                velocity.x += x_result;
-                velocity.y += y_result;
-                velocity.z += z_result;
-            }
-
-            angular.x += velocity.x;
-            angular.y += velocity.y;
-            angular.z += velocity.z;
-        }
-    }
-}
+pub mod controller;
+pub mod keyboard;
+pub mod mouse;
